+++ src/main/ts/args/Dispatch.ts
@@ -3,6 +3,7 @@ import * as Release from '../commands/Release';
 import * as Advance from '../commands/Advance';
 import * as Stamp from '../commands/Stamp';
 import * as Publish from '../commands/Publish';
+import * as Revive from '../commands/Revive';
 import * as Status from '../commands/Status';
 import * as BeehiveArgs from './BeehiveArgs';
 
@@ -17,6 +18,7 @@ export const dispatch = (args: BeehiveArgs): Promise<void> =>
     Advance.advanceCi,
     Stamp.stamp,
     Publish.publish,
-    Status.status
+    Status.status,
+    Revive.revive
   );