+++ src/main/ts/commands/Release.ts
@@ -2,6 +2,7 @@ import * as O from 'fp-ts/Option';
 import { ReleaseArgs } from '../args/BeehiveArgs';
 import * as Version from '../core/Version';
 import * as Git from '../utils/Git';
+import * as AdvanceVersion from '../core/AdvanceVersion';
 import * as PackageJson from '../core/PackageJson';
 import * as PromiseUtils from '../utils/PromiseUtils';
 import { BranchState, getBranchDetails } from '../core/BranchLogic';
@@ -9,12 +10,6 @@ import { printHeaderMessage } from '../core/Messages';
 
 type Version = Version.Version;
 const { versionToString } = Version;
-
-export const updateVersion = (version: Version): Version => ({
-  major: version.major,
-  minor: version.minor,
-  patch: version.patch
-});
 
 export const release = async (args: ReleaseArgs): Promise<void> => {
   printHeaderMessage(args);
@@ -38,7 +33,7 @@ export const release = async (args: ReleaseArgs): Promise<void> => {
     return PromiseUtils.fail('Branch is not in Release Candidate state - can\'t release.');
   }
 
-  const newVersion = updateVersion(branchDetails.version);
+  const newVersion = AdvanceVersion.updateToStable(branchDetails.version);
   console.log(`Updating version from ${versionToString(branchDetails.version)} to ${versionToString(newVersion)}`);
 
   const rootModule = branchDetails.rootModule;